assets/snowflake_cortex_asset/requirements.txt
dagster>=1.8.0
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.0
